v159-the maximum size of file can be set +some fixes

+++ back-end/src/middlewares/error.js
@@ -4,9 +4,15 @@ const winston = require("winston");
 const deleteFile = require("../helpers/deleteFile");
 
 module.exports = (err, req, res, next) => {
-  if (req.file)
+  if (req.file && req.file.path)
     //if some files uploaded with this req , delete them
     deleteFile(req.file.path, req.file.path);
+
+  if (req.files && req.files.length) {
+    for (let file of req.files) {
+      file.path ? deleteFile(file.path, file.path) : null;
+    }
+  }
   winston.error(err.message, err);
   res.status(500).json({ message: "خطای سرور" });
 };
