+++ back-end/src/routes/admin/controller.js
@@ -30,7 +30,7 @@ module.exports = new (class extends controller {
       });
 
     if (result.status === 404) {
-      if (req.file)
+      if (req.file && req.file.path)
         //if some files uploaded with this req , delete them
         deleteFile(req.file.path, req.file.path);
 
@@ -270,7 +270,7 @@ module.exports = new (class extends controller {
       });
 
     if (result.status === 404) {
-      if (req.file)
+      if (req.file && req.file.path)
         //if some files uploaded with this req , delete them
         deleteFile(req.file.path, req.file.path);
